update to 1.21
