+++ gradle.properties
@@ -4,14 +4,14 @@ org.gradle.parallel=true
 
 # Fabric Properties
 # check these on https://fabricmc.net/develop
-minecraft_version=1.20.4
-yarn_mappings=1.20.4+build.3
-loader_version=0.15.7
+minecraft_version=1.21
+yarn_mappings=1.21+build.1
+loader_version=0.15.11
 
 # Mod Properties
 mod_version=1.0.0
-maven_group=com.kimlan.saveanimals
-archives_base_name=save-animals
+maven_group=com.kimlan.protectanimals
+archives_base_name=protect-animals
 
 # Dependencies
-fabric_version=0.96.4+1.20.4
+fabric_version=0.100.1+1.21
