+++ src/main/java/com/kimlan/protectanimals/ProtectAnimals.java
@@ -1,15 +1,15 @@
-package com.kimlan.saveanimals;
+package com.kimlan.protectanimals;
 
 import net.fabricmc.api.ModInitializer;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class SaveAnimals implements ModInitializer {
+public class ProtectAnimals implements ModInitializer {
 	// This logger is used to write text to the console and the log file.
 	// It is considered best practice to use your mod id as the logger's name.
 	// That way, it's clear which mod wrote info, warnings, and errors.
-    public static final Logger LOGGER = LoggerFactory.getLogger("save-animals");
+    public static final Logger LOGGER = LoggerFactory.getLogger("protect-animals");
 
 	@Override
 	public void onInitialize() {
