+++ src/main/java/com/kimlan/protectanimals/mixin/AnimalEntityMixin.java
@@ -1,16 +1,16 @@
-package com.kimlan.saveanimals.mixin;
+package com.kimlan.protectanimals.mixin;
 
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.damage.DamageSource;
 import net.minecraft.entity.mob.Monster;
-import net.minecraft.entity.passive.AllayEntity;
+import net.minecraft.entity.passive.AnimalEntity;
 import net.minecraft.world.World;
 import org.spongepowered.asm.mixin.Mixin;
 
-@Mixin(AllayEntity.class)
-abstract class AllayEntityMixin extends Entity {
-    public AllayEntityMixin(EntityType<?> type, World world) {
+@Mixin(AnimalEntity.class)
+abstract class AnimalEntityMixin extends Entity {
+    public AnimalEntityMixin(EntityType<?> type, World world) {
         super(type, world);
     }
