+++ src/main/resources/assets/protect-animals/icon.png
