+++ src/main/resources/protect-animals.mixins.json
@@ -1,7 +1,7 @@
 {
 	"required": true,
-	"package": "com.kimlan.saveanimals.mixin",
-	"compatibilityLevel": "JAVA_17",
+	"package": "com.kimlan.protectanimals.mixin",
+	"compatibilityLevel": "JAVA_21",
 	"mixins": [
 		"AllayEntityMixin",
 		"AnimalEntityMixin",
